requirements.txt
# Основные зависимости для PyTorch
torch>=2.0.0
torchvision>=0.15.0

# YOLOv8 для Object Detection
ultralytics>=8.0.0

# Обработка изображений
Pillow>=9.0.0
opencv-python>=4.5.0

# FastAPI и веб-сервер
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Научные вычисления
numpy>=1.21.0
pandas>=1.3.0

# Визуализация и логирование
matplotlib>=3.5.0
seaborn>=0.11.0

# Утилиты
tqdm>=4.64.0
pathlib2>=2.3.0

# Дополнительные зависимости для YOLO
pyyaml>=6.0
requests>=2.25.0

# Дополнительные зависимости для разработки (опционально)
# jupyter>=1.0.0
# ipykernel>=6.0.0
# black>=22.0.0
# flake8>=4.0.0
